requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Uploads==0.1.3
Jinja2==2.7.2
MarkupSafe==0.19
Pillow==2.4.0
Werkzeug==0.9.4
itsdangerous==0.23
pyhashxx==0.1.1
redis==2.9.1
requests==2.2.1
wsgiref==0.1.2